requirements.txt
boto3
anthropic
python-dotenv
redis
fastapi
uvicorn
websockets
psutil
requests
elasticsearch
psycopg2-binary
